separa o código da modal da page edital
